refactor: changed function structure
